change field types integer-numeric
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/gwidgets@364 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/gWidgetstcltk/R/aaaR5classes.R
@@ -338,8 +338,8 @@ setRefClass("Entry",
             fields=list(
               m="tkwin",
               l="tkwin",
-              lindex = "integer",  # index of selection widget
-              no.wds = "integer",  # track number of possible wds to choose from
+              lindex = "numeric",  # index of selection widget
+              no.wds = "numeric",  # track number of possible wds to choose from
               words = "character",
               max.words = "numeric", # maximum words in a display
               init_msg = "character" # an initial message
